fix(ExecutionContext): changes signature to write 4 bytes instead of 8
Prior, we passed a 'size_t' to allocateVariableSizedData(), which we used to write to the arena (8 bytes). However, we expected to write 4 bytes. This could lead to a scenario in which we overwrote bytes of the next buffer, e.g., bytes of the control block.

+++ nes-runtime/include/ExecutionContext.hpp
@@ -71,7 +71,7 @@ struct ArenaRef
     /// Allocates memory from the arena. If the available space for the pointer is smaller than the required size, we allocate a new buffer from the arena.
     nautilus::val<int8_t*> allocateMemory(const nautilus::val<size_t>& sizeInBytes);
 
-    VariableSizedData allocateVariableSizedData(const nautilus::val<size_t>& sizeInBytes);
+    VariableSizedData allocateVariableSizedData(const nautilus::val<uint32_t>& sizeInBytes);
 
 private:
     nautilus::val<Arena*> arenaRef;
